feat: use subject for report name
Following internal discussions, we use the subject line for both the
google drive folder structure and the document name.  For example, if
the subject line is "HZ 123 - My Report 2019" and an MS Excel document
is attached, the Google Drive folder will be "hz 123" and the file name
on Google Drive will be "my report 2019.xlsx".